HTBHF-2377 Refactored DetermineEntitlementMessageProcessor ineligible decision logic into new class: IneligibleEntitlementDecisionHandler
