+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/message/processor/DetermineEntitlementMessageProcessorTest.java
@@ -2,18 +2,10 @@ package uk.gov.dhsc.htbhf.claimant.message.processor;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
-import org.mockito.ArgumentCaptor;
-import org.mockito.InOrder;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.dhsc.htbhf.claimant.communications.DetermineEntitlementNotificationHandler;
-import uk.gov.dhsc.htbhf.claimant.entitlement.PregnancyEntitlementCalculator;
-import uk.gov.dhsc.htbhf.claimant.entity.CardStatus;
+import uk.gov.dhsc.htbhf.claimant.entitlement.IneligibleEntitlementDecisionHandler;
 import uk.gov.dhsc.htbhf.claimant.entity.Claim;
 import uk.gov.dhsc.htbhf.claimant.entity.Message;
 import uk.gov.dhsc.htbhf.claimant.entity.PaymentCycle;
@@ -24,42 +16,28 @@ import uk.gov.dhsc.htbhf.claimant.message.MessageType;
 import uk.gov.dhsc.htbhf.claimant.message.context.DetermineEntitlementMessageContext;
 import uk.gov.dhsc.htbhf.claimant.message.context.MessageContextLoader;
 import uk.gov.dhsc.htbhf.claimant.message.payload.MessagePayload;
-import uk.gov.dhsc.htbhf.claimant.model.ClaimStatus;
 import uk.gov.dhsc.htbhf.claimant.model.eligibility.EligibilityAndEntitlementDecision;
-import uk.gov.dhsc.htbhf.claimant.model.eligibility.QualifyingBenefitEligibilityStatus;
-import uk.gov.dhsc.htbhf.claimant.repository.ClaimRepository;
-import uk.gov.dhsc.htbhf.claimant.service.ClaimMessageSender;
 import uk.gov.dhsc.htbhf.claimant.service.EligibilityAndEntitlementService;
-import uk.gov.dhsc.htbhf.claimant.service.audit.EventAuditor;
 import uk.gov.dhsc.htbhf.claimant.service.payments.PaymentCycleService;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.inOrder;
-import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyZeroInteractions;
-import static org.mockito.internal.verification.VerificationModeFactory.times;
-import static uk.gov.dhsc.htbhf.claimant.entity.CardStatus.PENDING_CANCELLATION;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static uk.gov.dhsc.htbhf.claimant.message.MessageStatus.COMPLETED;
 import static uk.gov.dhsc.htbhf.claimant.message.MessageType.DETERMINE_ENTITLEMENT;
-import static uk.gov.dhsc.htbhf.claimant.model.eligibility.QualifyingBenefitEligibilityStatus.CONFIRMED;
-import static uk.gov.dhsc.htbhf.claimant.model.eligibility.QualifyingBenefitEligibilityStatus.NOT_CONFIRMED;
-import static uk.gov.dhsc.htbhf.claimant.reporting.ClaimAction.UPDATED_FROM_ACTIVE_TO_EXPIRED;
-import static uk.gov.dhsc.htbhf.claimant.reporting.ClaimAction.UPDATED_FROM_ACTIVE_TO_PENDING_EXPIRY;
 import static uk.gov.dhsc.htbhf.claimant.testsupport.ClaimTestDataFactory.aClaimWithExpectedDeliveryDateAndChildrenDobs;
 import static uk.gov.dhsc.htbhf.claimant.testsupport.EligibilityAndEntitlementTestDataFactory.aDecisionWithStatus;
-import static uk.gov.dhsc.htbhf.claimant.testsupport.EligibilityAndEntitlementTestDataFactory.aDecisionWithStatusAndChildren;
 import static uk.gov.dhsc.htbhf.claimant.testsupport.MessageContextTestDataFactory.aDetermineEntitlementMessageContext;
 import static uk.gov.dhsc.htbhf.claimant.testsupport.MessageTestDataFactory.aValidMessageWithType;
 import static uk.gov.dhsc.htbhf.claimant.testsupport.PaymentCycleTestDataFactory.aPaymentCycleWithClaim;
 import static uk.gov.dhsc.htbhf.claimant.testsupport.PaymentCycleTestDataFactory.aPaymentCycleWithStartDateAndClaim;
-import static uk.gov.dhsc.htbhf.claimant.testsupport.TestConstants.*;
+import static uk.gov.dhsc.htbhf.claimant.testsupport.TestConstants.EXPECTED_DELIVERY_DATE_IN_TWO_MONTHS;
+import static uk.gov.dhsc.htbhf.claimant.testsupport.TestConstants.ONE_CHILD_UNDER_ONE_AND_ONE_CHILD_BETWEEN_ONE_AND_FOUR;
 import static uk.gov.dhsc.htbhf.eligibility.model.EligibilityStatus.ELIGIBLE;
 import static uk.gov.dhsc.htbhf.eligibility.model.EligibilityStatus.INELIGIBLE;
 
@@ -73,19 +51,9 @@ class DetermineEntitlementMessageProcessorTest {
     @Mock
     private PaymentCycleService paymentCycleService;
     @Mock
-    private ClaimRepository claimRepository;
-    @Mock
     private MessageQueueClient messageQueueClient;
     @Mock
-    private DetermineEntitlementNotificationHandler determineEntitlementNotificationHandler;
-    @Mock
-    private PregnancyEntitlementCalculator pregnancyEntitlementCalculator;
-    @Mock
-    private ChildDateOfBirthCalculator childDateOfBirthCalculator;
-    @Mock
-    private EventAuditor eventAuditor;
-    @Mock
-    private ClaimMessageSender claimMessageSender;
+    private IneligibleEntitlementDecisionHandler ineligibleEntitlementDecisionHandler;
 
     @InjectMocks
     private DetermineEntitlementMessageProcessor processor;
@@ -120,150 +88,20 @@ class DetermineEntitlementMessageProcessorTest {
         verify(paymentCycleService).updatePaymentCycle(context.getCurrentPaymentCycle(), decision);
         MessagePayload expectedPayload = MessagePayloadFactory.buildMakePaymentMessagePayload(context.getCurrentPaymentCycle());
         verify(messageQueueClient).sendMessage(expectedPayload, MessageType.MAKE_PAYMENT);
-        verifyZeroInteractions(claimRepository, pregnancyEntitlementCalculator, childDateOfBirthCalculator,
-                childDateOfBirthCalculator, eventAuditor, claimMessageSender);
+        verifyNoMoreInteractions(ineligibleEntitlementDecisionHandler);
     }
 
-    @ParameterizedTest(name = "Qualifying benefit status={0}")
-    @ValueSource(strings = {"CONFIRMED", "NOT_CONFIRMED"})
-    void shouldExpireClaimWhenIneligibleWithNoChildrenAndNotPregnant(QualifyingBenefitEligibilityStatus qualifyingBenefitEligibilityStatus) {
+    @Test
+    void shouldSuccessfullyProcessMessageAndHandleIneligibleDecision() {
         //Given
-        //Test for HTBHF-2182 has the following context:
-        // Previous cycle: children exist and are under 4 but will be 4 in the next cycle, not pregnant
-        // Current cycle: no children and not pregnant
-        List<LocalDate> currentCycleChildrenDobs = NO_CHILDREN;
-        DetermineEntitlementMessageContext context = buildMessageContext(
-                NOT_PREGNANT,
-                SINGLE_NEARLY_FOUR_YEAR_OLD,
-                NOT_PREGNANT,
-                currentCycleChildrenDobs);
-        given(messageContextLoader.loadDetermineEntitlementContext(any())).willReturn(context);
-
-        //Claimant has children under 4 at the previous payment cycle that would still over 4 now (but not reported by DWP).
-        given(childDateOfBirthCalculator.hadChildrenUnder4AtStartOfPaymentCycle(any())).willReturn(true);
-        given(childDateOfBirthCalculator.hadChildrenUnderFourAtGivenDate(any(), any())).willReturn(false);
-
-        //Eligibility is INELIGIBLE as they have no children or pregnancy but did in the previous cycle. DWP status is irrelevant here.
-        EligibilityAndEntitlementDecision decision = aDecisionWithStatusAndChildren(INELIGIBLE, qualifyingBenefitEligibilityStatus, currentCycleChildrenDobs);
-        given(eligibilityAndEntitlementService.evaluateExistingClaimant(any(), any(), any())).willReturn(decision);
-        given(pregnancyEntitlementCalculator.isEntitledToVoucher(any(), any())).willReturn(false);
-
-        //Current payment cycle voucher entitlement mocking
-        Message message = aValidMessageWithType(DETERMINE_ENTITLEMENT);
-
-        //When
-        MessageStatus messageStatus = processor.processMessage(message);
-
-        //Then
-        runCommonVerifications(context, decision, message, messageStatus, ClaimStatus.EXPIRED, PENDING_CANCELLATION);
-        LocalDate currentPaymentCycleStartDate = context.getCurrentPaymentCycle().getCycleStartDate();
-        verify(pregnancyEntitlementCalculator).isEntitledToVoucher(NOT_PREGNANT, currentPaymentCycleStartDate);
-        verify(childDateOfBirthCalculator).hadChildrenUnder4AtStartOfPaymentCycle(context.getPreviousPaymentCycle());
-        verify(childDateOfBirthCalculator).hadChildrenUnderFourAtGivenDate(SINGLE_NEARLY_FOUR_YEAR_OLD, currentPaymentCycleStartDate);
-        verify(eventAuditor).auditExpiredClaim(context.getClaim());
-        verify(claimMessageSender).sendReportClaimMessage(context.getClaim(), currentCycleChildrenDobs, UPDATED_FROM_ACTIVE_TO_EXPIRED);
-        verifyZeroInteractions(messageQueueClient, determineEntitlementNotificationHandler);
-    }
-
-    @ParameterizedTest(name = "Expected delivery date previous cycle={0}")
-    @MethodSource("provideArgumentsForPreviousCycleExpectedDeliveryDate")
-    void shouldExpireClaimWhenChildDisappearsFromFeedAndNotPregnant(LocalDate previousCycleExpectedDeliveryDate) {
-        //Given
-        //Test for HTBHF-2185 has the following context:
-        // Previous cycle: children exist and are under 4, not pregnant or pregnant (parameterised)
-        // Current cycle: no children and not pregnant
-        List<LocalDate> currentCycleChildrenDobs = NO_CHILDREN;
-        DetermineEntitlementMessageContext context = buildMessageContext(
-                previousCycleExpectedDeliveryDate,
-                SINGLE_THREE_YEAR_OLD,
-                NOT_PREGNANT,
-                currentCycleChildrenDobs);
-        given(messageContextLoader.loadDetermineEntitlementContext(any())).willReturn(context);
-
-        //Claimant has children under 4 at the previous payment cycle that would still be under 4 now (but not reported by DWP).
-        given(childDateOfBirthCalculator.hadChildrenUnder4AtStartOfPaymentCycle(any())).willReturn(true);
-        given(childDateOfBirthCalculator.hadChildrenUnderFourAtGivenDate(any(), any())).willReturn(true);
-
-        //Eligibility is INELIGIBLE as they have no children or pregnancy but DWP have returned that they are eligible (CONFIRMED)
-        EligibilityAndEntitlementDecision decision = aDecisionWithStatusAndChildren(INELIGIBLE, CONFIRMED, currentCycleChildrenDobs);
-        given(eligibilityAndEntitlementService.evaluateExistingClaimant(any(), any(), any())).willReturn(decision);
-        given(pregnancyEntitlementCalculator.isEntitledToVoucher(any(), any())).willReturn(false);
-
-        //Current payment cycle voucher entitlement mocking
-        Message message = aValidMessageWithType(DETERMINE_ENTITLEMENT);
-
-        //When
-        MessageStatus messageStatus = processor.processMessage(message);
-
-        //Then
-        runCommonVerifications(context, decision, message, messageStatus, ClaimStatus.EXPIRED, PENDING_CANCELLATION);
-        verify(determineEntitlementNotificationHandler).sendNoChildrenOnFeedClaimNoLongerEligibleEmail(context.getClaim());
-        LocalDate currentPaymentCycleStartDate = context.getCurrentPaymentCycle().getCycleStartDate();
-        verify(pregnancyEntitlementCalculator).isEntitledToVoucher(NOT_PREGNANT, currentPaymentCycleStartDate);
-        verify(childDateOfBirthCalculator).hadChildrenUnder4AtStartOfPaymentCycle(context.getPreviousPaymentCycle());
-        verify(childDateOfBirthCalculator).hadChildrenUnderFourAtGivenDate(SINGLE_THREE_YEAR_OLD, currentPaymentCycleStartDate);
-        verify(eventAuditor).auditExpiredClaim(context.getClaim());
-        verify(claimMessageSender).sendReportClaimMessage(context.getClaim(), currentCycleChildrenDobs, UPDATED_FROM_ACTIVE_TO_EXPIRED);
-        verifyZeroInteractions(messageQueueClient);
-    }
-
-    //HTBHF-1757 Children in current cycle, DWP returns ineligible, varying on pregnancy and children in previous cycle
-    @ParameterizedTest(name = "Children DOB previous cycle={0}, expected delivery date previous cycle={1}, expected delivery date current cycle={2}")
-    @MethodSource("provideArgumentsForPendingExpiryTestsWithChildrenInCurrentCycle")
-    void shouldUpdateClaimToPendingExpiryWhenClaimantIsNotEligibleWithChildrenInCurrentCycle(
-            List<LocalDate> previousCycleChildrenDobs,
-            LocalDate previousCycleExpectedDeliveryDate,
-            LocalDate currentCycleExpectedDeliveryDate) {
-        //Given - their children would still be under 4 but DWP doesn't return them if they're NOT_CONFIRMED.
-        List<LocalDate> currentCycleChildrenDobs = NO_CHILDREN;
-        DetermineEntitlementMessageContext context = buildMessageContext(
-                previousCycleExpectedDeliveryDate,
-                previousCycleChildrenDobs,
-                currentCycleExpectedDeliveryDate,
-                currentCycleChildrenDobs);
-        given(messageContextLoader.loadDetermineEntitlementContext(any())).willReturn(context);
-
-        //Eligibility - any status that isn't ELIGIBLE has the same effect here.
-        EligibilityAndEntitlementDecision decision = aDecisionWithStatusAndChildren(INELIGIBLE, NOT_CONFIRMED, currentCycleChildrenDobs);
-        given(eligibilityAndEntitlementService.evaluateExistingClaimant(any(), any(), any())).willReturn(decision);
-        given(pregnancyEntitlementCalculator.isEntitledToVoucher(any(), any())).willReturn(false);
-        //Had children in last cycle who are under 4 at the start of the current payment cycle.
-        given(childDateOfBirthCalculator.hadChildrenUnder4AtStartOfPaymentCycle(any())).willReturn(true);
-        given(childDateOfBirthCalculator.hadChildrenUnderFourAtGivenDate(any(), any())).willReturn(true);
-
-        //Current payment cycle voucher entitlement mocking
-        Message message = aValidMessageWithType(DETERMINE_ENTITLEMENT);
-
-        //When
-        MessageStatus messageStatus = processor.processMessage(message);
-
-        //Then
-        runCommonVerifications(context, decision, message, messageStatus, ClaimStatus.PENDING_EXPIRY, PENDING_CANCELLATION);
-        verify(determineEntitlementNotificationHandler).sendClaimNoLongerEligibleEmail(context.getClaim());
-        LocalDate currentCycleStartDate = context.getCurrentPaymentCycle().getCycleStartDate();
-        verify(childDateOfBirthCalculator).hadChildrenUnder4AtStartOfPaymentCycle(context.getPreviousPaymentCycle());
-        verify(childDateOfBirthCalculator).hadChildrenUnderFourAtGivenDate(ONE_CHILD_UNDER_ONE_AND_ONE_CHILD_BETWEEN_ONE_AND_FOUR, currentCycleStartDate);
-        verify(pregnancyEntitlementCalculator).isEntitledToVoucher(currentCycleExpectedDeliveryDate, currentCycleStartDate);
-        verify(claimMessageSender).sendReportClaimMessage(context.getClaim(), currentCycleChildrenDobs, UPDATED_FROM_ACTIVE_TO_PENDING_EXPIRY);
-        verifyZeroInteractions(childDateOfBirthCalculator, eventAuditor);
-    }
-
-    //HTBHF-1757 No children in current cycle, are pregnant but ineligible from DWP
-    @ParameterizedTest(name = "Children DOB in previous cycle={0}")
-    @MethodSource("provideArgumentsForChildrenInPreviousCycle")
-    void shouldUpdateClaimToPendingExpiryWhenClaimantIsNotEligibleButStillPregnant(List<LocalDate> previousCycleChildrenDobs) {
-        //Given
-        List<LocalDate> currentCycleChildrenDobs = NO_CHILDREN;
         DetermineEntitlementMessageContext context = buildMessageContext(
                 EXPECTED_DELIVERY_DATE_IN_TWO_MONTHS,
-                previousCycleChildrenDobs,
+                ONE_CHILD_UNDER_ONE_AND_ONE_CHILD_BETWEEN_ONE_AND_FOUR,
                 EXPECTED_DELIVERY_DATE_IN_TWO_MONTHS,
-                currentCycleChildrenDobs);
+                ONE_CHILD_UNDER_ONE_AND_ONE_CHILD_BETWEEN_ONE_AND_FOUR);
         given(messageContextLoader.loadDetermineEntitlementContext(any())).willReturn(context);
-        given(pregnancyEntitlementCalculator.isEntitledToVoucher(any(), any())).willReturn(true);
 
-        //Eligibility - any status that isn't ELIGIBLE has the same effect here.
-        EligibilityAndEntitlementDecision decision = aDecisionWithStatusAndChildren(INELIGIBLE, NOT_CONFIRMED, currentCycleChildrenDobs);
+        EligibilityAndEntitlementDecision decision = aDecisionWithStatus(INELIGIBLE);
         given(eligibilityAndEntitlementService.evaluateExistingClaimant(any(), any(), any())).willReturn(decision);
 
         //Current payment cycle voucher entitlement mocking
@@ -273,63 +111,6 @@ class DetermineEntitlementMessageProcessorTest {
         MessageStatus messageStatus = processor.processMessage(message);
 
         //Then
-        runCommonVerifications(context, decision, message, messageStatus, ClaimStatus.PENDING_EXPIRY, PENDING_CANCELLATION);
-        verify(determineEntitlementNotificationHandler).sendClaimNoLongerEligibleEmail(context.getClaim());
-        verify(pregnancyEntitlementCalculator).isEntitledToVoucher(EXPECTED_DELIVERY_DATE_IN_TWO_MONTHS, context.getCurrentPaymentCycle().getCycleStartDate());
-        verify(claimMessageSender).sendReportClaimMessage(context.getClaim(), currentCycleChildrenDobs, UPDATED_FROM_ACTIVE_TO_PENDING_EXPIRY);
-        verifyZeroInteractions(childDateOfBirthCalculator, eventAuditor);
-    }
-
-    @ParameterizedTest(name = "Qualifying benefit status={0}")
-    @ValueSource(strings = {"CONFIRMED", "NOT_CONFIRMED"})
-    void shouldExpireClaimWhenClaimantWasPregnantWithNoChildrenButNoLongerPregnant(QualifyingBenefitEligibilityStatus qualifyingBenefitEligibilityStatus) {
-        //Given
-        List<LocalDate> currentCycleChildrenDobs = NO_CHILDREN;
-        DetermineEntitlementMessageContext context = buildMessageContext(
-                EXPECTED_DELIVERY_DATE_IN_TWO_MONTHS,
-                NO_CHILDREN,
-                NOT_PREGNANT,
-                currentCycleChildrenDobs);
-        given(messageContextLoader.loadDetermineEntitlementContext(any())).willReturn(context);
-        LocalDate currentPaymentCycleStartDate = context.getCurrentPaymentCycle().getCycleStartDate();
-        LocalDate previousPaymentCycleStartDate = context.getPreviousPaymentCycle().getCycleStartDate();
-        lenient().when(pregnancyEntitlementCalculator.isEntitledToVoucher(NOT_PREGNANT, currentPaymentCycleStartDate)).thenReturn(false);
-        lenient().when(pregnancyEntitlementCalculator.isEntitledToVoucher(EXPECTED_DELIVERY_DATE_IN_TWO_MONTHS, previousPaymentCycleStartDate))
-                .thenReturn(true);
-        //Had children in last cycle who are under 4 at the start of the current payment cycle.
-        given(childDateOfBirthCalculator.hadChildrenUnder4AtStartOfPaymentCycle(any())).willReturn(false);
-
-        //Eligibility - any status that isn't ELIGIBLE has the same effect here.
-        EligibilityAndEntitlementDecision decision = aDecisionWithStatusAndChildren(INELIGIBLE, qualifyingBenefitEligibilityStatus, currentCycleChildrenDobs);
-        given(eligibilityAndEntitlementService.evaluateExistingClaimant(any(), any(), any())).willReturn(decision);
-
-        //Current payment cycle voucher entitlement mocking
-        Message message = aValidMessageWithType(DETERMINE_ENTITLEMENT);
-
-        //When
-        MessageStatus messageStatus = processor.processMessage(message);
-
-        //Then
-        runCommonVerifications(context, decision, message, messageStatus, ClaimStatus.EXPIRED, PENDING_CANCELLATION);
-        verify(childDateOfBirthCalculator).hadChildrenUnder4AtStartOfPaymentCycle(context.getPreviousPaymentCycle());
-        InOrder inOrder = inOrder(pregnancyEntitlementCalculator, pregnancyEntitlementCalculator);
-        inOrder.verify(pregnancyEntitlementCalculator).isEntitledToVoucher(NOT_PREGNANT, currentPaymentCycleStartDate);
-        inOrder.verify(pregnancyEntitlementCalculator).isEntitledToVoucher(EXPECTED_DELIVERY_DATE_IN_TWO_MONTHS, previousPaymentCycleStartDate);
-        verify(eventAuditor).auditExpiredClaim(context.getClaim());
-        verify(claimMessageSender).sendReportClaimMessage(context.getClaim(), currentCycleChildrenDobs, UPDATED_FROM_ACTIVE_TO_EXPIRED);
-        verifyZeroInteractions(determineEntitlementNotificationHandler, messageQueueClient);
-    }
-
-    private void verifyClaimSavedWithStatus(ClaimStatus claimStatus, CardStatus cardStatus) {
-        ArgumentCaptor<Claim> argumentCaptor = ArgumentCaptor.forClass(Claim.class);
-        verify(claimRepository, times(2)).save(argumentCaptor.capture());
-        List<Claim> claims = argumentCaptor.getAllValues();
-        assertThat(claims.get(0).getClaimStatus()).isEqualTo(claimStatus);
-        assertThat(claims.get(1).getCardStatus()).isEqualTo(cardStatus);
-    }
-
-    private void runCommonVerifications(DetermineEntitlementMessageContext context, EligibilityAndEntitlementDecision decision,
-                                        Message message, MessageStatus messageStatus, ClaimStatus expectedClaimStatus, CardStatus expectedCardStatus) {
         assertThat(messageStatus).isEqualTo(COMPLETED);
         verify(messageContextLoader).loadDetermineEntitlementContext(message);
         verify(eligibilityAndEntitlementService).evaluateExistingClaimant(context.getClaim().getClaimant(),
@@ -337,7 +118,9 @@ class DetermineEntitlementMessageProcessorTest {
                 context.getPreviousPaymentCycle());
 
         verify(paymentCycleService).updatePaymentCycle(context.getCurrentPaymentCycle(), decision);
-        verifyClaimSavedWithStatus(expectedClaimStatus, expectedCardStatus);
+        verify(ineligibleEntitlementDecisionHandler)
+                .handleIneligibleDecision(context.getClaim(), context.getPreviousPaymentCycle(), context.getCurrentPaymentCycle(), decision);
+        verifyNoMoreInteractions(messageQueueClient);
     }
 
     private DetermineEntitlementMessageContext buildMessageContext(LocalDate previousCycleExpectedDeliveryDate,
@@ -360,28 +143,6 @@ class DetermineEntitlementMessageProcessorTest {
                 currentPaymentCycle,
                 previousPaymentCycle,
                 claimAtCurrentCycle);
-    }
-
-    //Argument order is: previousCycleChildrenDobs, previousCycleExpectedDeliveryDate, currentCycleExpectedDeliveryDate
-    private static Stream<Arguments> provideArgumentsForPendingExpiryTestsWithChildrenInCurrentCycle() {
-        return Stream.of(
-                Arguments.of(ONE_CHILD_UNDER_ONE_AND_ONE_CHILD_BETWEEN_ONE_AND_FOUR, NOT_PREGNANT, NOT_PREGNANT),
-                Arguments.of(ONE_CHILD_UNDER_ONE_AND_ONE_CHILD_BETWEEN_ONE_AND_FOUR, EXPECTED_DELIVERY_DATE_IN_TWO_MONTHS, EXPECTED_DELIVERY_DATE_IN_TWO_MONTHS)
-        );
-    }
-
-    private static Stream<Arguments> provideArgumentsForChildrenInPreviousCycle() {
-        return Stream.of(
-                Arguments.of(ONE_CHILD_UNDER_ONE_AND_ONE_CHILD_BETWEEN_ONE_AND_FOUR),
-                Arguments.of(NO_CHILDREN)
-        );
-    }
-
-    private static Stream<Arguments> provideArgumentsForPreviousCycleExpectedDeliveryDate() {
-        return Stream.of(
-                Arguments.of(NOT_PREGNANT),
-                Arguments.of(EXPECTED_DELIVERY_DATE_IN_TWO_MONTHS)
-        );
     }
 
 }
